+++ frontend/src/pages/explore/HashesView.tsx
@@ -1,0 +1,5 @@
+import { DataExplorerView } from './DataExplorerView'
+
+export function HashesView() {
+  return <DataExplorerView domain="hashes" />
+}